pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pepoapple-core"
version = "0.1.0"
description = "Pepoapple Proxy Manager + Billing MVP"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.5.2",
  "python-multipart>=0.0.9",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "redis>=5.2.0",
  "strawberry-graphql[fastapi]>=0.279.0",
  "httpx>=0.27.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "httpx>=0.27.2",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]
